Migrate Mutation.updateBaseSchema
